+++ collector/src/test/java/com/bitmovin/analytics/data/SimpleEventDataDispatcherTest.kt
@@ -17,7 +17,6 @@ import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import org.assertj.core.api.Assertions
 import org.junit.After
 import org.junit.Before
 import org.junit.Ignore
@@ -65,36 +64,7 @@ class SimpleEventDataDispatcherTest {
         verify { backendFactory.createBackend(config, context, any()) }
     }
 
-    @Test
-    fun `sequence numbers auto-increments on add`() {
-        dispatcher.enable()
-        val eventData1 = createTestEventData()
-        dispatcher.add(eventData1)
-        val eventData2 = createTestEventData()
-        dispatcher.add(eventData2)
-        Assertions.assertThat(eventData1.sequenceNumber).isEqualTo(eventData2.sequenceNumber - 1)
-    }
-
-    @Test
-    fun `sequence number is limited to max-limit`() {
-        dispatcher.enable()
-
-        // enable the backend to forward the data
-        dispatcher.authenticationCompleted(
-            AuthenticationResponse.Granted("authenticated-key", null),
-        )
-
-        val eventData = createTestEventData()
-        for (i in 0..1005) {
-            dispatcher.add(eventData)
-        }
-        Assertions.assertThat(eventData.sequenceNumber).isEqualTo(1000)
-
-        // this line might cause issues when there are too many calls (verification would throw an error)
-        // test is stalling then. in case of the correct amount of calls, test passes
-        verify(exactly = 1001) { backend.send(any()) }
-    }
-
+    // FIXME: why do we need the scope in the constructor?
     @Test
     @Ignore("This test fails as the scope created in the constructor is not cancelled. This is a bug.")
     fun `disabling the dispatcher cancels the scope`() {